Renaming 'mock_server' directory to 'mockServer'
