+++ mifos-passcode-cmp/src/desktopMain/kotlin/auth/deviceAuth/windows/WindowsHelloAuthenticatorNativeSupport.kt
@@ -1,6 +1,9 @@
 package auth.deviceAuth.windows
 
-import com.mifos.passcode.mock_server.*
+import com.mifos.passcode.mockServer.RegistrationDataGET
+import com.mifos.passcode.mockServer.RegistrationDataPOST
+import com.mifos.passcode.mockServer.VerificationDataGET
+import com.mifos.passcode.mockServer.VerificationDataPOST
 import com.sun.jna.Library
 
 /**
